+++ packages/maps/src/lib/index.js
@@ -30,6 +30,12 @@ export { default as MapMarkerFlyToFeature } from './mapMarker/elements/mapMarker
 export { default as MapMarkerPlacement } from './mapMarker/elements/mapMarkerPlacement/MapMarkerPlacement.svelte';
 export { default as MapMarkerStyledContainer } from './mapMarker/elements/mapMarkerStyledContainer/MapMarkerStyledContainer.svelte';
 
+export { default as MapDeckPopovers } from './mapDeckPopovers/MapDeckPopovers.svelte';
+export { onClickPopoverHandler, clickedLayer, clickedFeature } from './mapDeckPopovers/stores';
+
+export { default as MapDeckTooltips } from './mapDeckTooltips/MapDeckTooltips.svelte';
+export { onMouseOverTooltipHandler, mousedOverObject } from './mapDeckTooltips/stores';
+
 export { default as MapPopover } from './mapPopover/MapPopover.svelte';
 
 // Layers