+++ packages/tables/src/lib/table/rows/headerRows/ColumnHeadingRow.svelte
@@ -51,7 +51,8 @@
 					style:width={col.computedWidth + 'px'}
 					aria-sort={colIsSortable ? order : ''}
 				>
-					<Header
+					<svelte:component
+						this={col.header?.renderer || Header}
 						label={col.label ?? col.short_label}
 						{order}
 						toggle={() => colIsSortable && table.toggleSort(col.short_label)}
@@ -61,6 +62,7 @@
 						hintText={col.hintText}
 						hintComponent={col.hintComponent}
 						hintType={col.hintType ?? 'tooltip'}
+						{...col.header || {}}
 					/>
 				</div>
 			{/if}